Move player related events
